requirements.txt
selenium~=3.141.0
bs4~=0.0.1
beautifulsoup4~=4.9.3
aiogram
python-dotenv
asyncio
asyncpg
aioschedule